Refactor job_control into 6 files and the same for the tests.
